\name{dynamicnetwork-package}
\alias{dynamicnetwork-package}
\docType{package}
\title{
Classes for Dynamic Relational Data
}
\description{
Tools to create and modify network objects that have edges and attributes changing in time.  The dynamic.network
class can represent a range of relational data types, and
supports arbitrary vertex/edge/graph attributes.
}
\details{


\tabular{ll}{
Package: \tab network\cr
Version: \tab 0.0-4\cr
Date: \tab August 25, 2006\cr
Depends: \tab R (>= 2.4.0),network\cr
Suggests: \tab sna,statnet\cr
License: \tab GPL version 2 (June, 1991)\cr
}


}
}
\author{
Skye Bender-deMoll <buttsc@uci.edu>, CSDE statnet team

Maintainer: Skye Bender-deMoll <skyebend@skyeome.net>
}
\keyword{ package }
